+++ yazi-binding/src/chord_cow.rs
@@ -1,7 +1,7 @@
 use mlua::UserData;
 use yazi_codegen::FromLuaOwned;
 
-#[derive(FromLuaOwned)]
+#[derive(Clone, FromLuaOwned)]
 pub struct ChordCow(pub yazi_config::keymap::ChordCow);
 
 impl From<ChordCow> for yazi_config::keymap::ChordCow {